manager session 넘기기

+++ src/main/java/com/thank/store/dao/ManagerDAO.java
@@ -18,7 +18,7 @@ public interface ManagerDAO {
 	
 	List<CvsProductDTO> getEnrolAvaiProductList(ManSearchDTO searchDTO) throws SQLException;
 
-	int getManagerNoFromMember(MemberDTO memberDTO) throws SQLException;
+	long getManagerNoFromMember(MemberDTO memberDTO) throws SQLException;
 
 	long getTotalRecord(ManSearchDTO searchDTO) throws SQLException;
 }
