+++ src/main/java/com/thank/store/service/ManagerService.java
@@ -20,7 +20,7 @@ public interface ManagerService {
 
 	ManPagingDTO getPagingInfo(ManSearchDTO searchDTO) throws Exception;
 
-	int getManagerNoFromMember(MemberDTO memberDTO) throws Exception;
+	long getManagerNoFromMember(MemberDTO memberDTO) throws Exception;
 	
 	long getDiscountPrice(long price, int discountRate) throws Exception;
